--- src/components/brødsmuler/Brødsmuler.less
@@ -1,27 +1,0 @@
-.brødsmuler {
-  font-size: 16px;
-  width: 100%;
-  margin-top: 1rem;
-  padding-bottom: 1rem;
-  text-align: center;
-  margin-right: 500px;
-
-  .brødsmule-element {
-    margin-left: 5px;
-    margin-right: 5px;
-
-    a {
-      text-decoration: none;
-    }
-
-    a:hover {
-      text-decoration: underline;
-    }
-  }
-
-  @media @tablet {
-    text-align: left;
-    margin-right: 0;
-    margin-left: 30px;
-  }
-}
