+++ src/components/brødsmuler/Brødsmuler.tsx
@@ -1,9 +1,38 @@
 import React from 'react';
-import { StyledBrødsmuler, Brødsmule } from './BrødsmulerElementer';
+import styled from 'styled-components';
+import { device } from '../../utils/styles';
+
+export const BrødsmulerWrapper = styled.div`
+  font-size: 16px;
+  width: 100%;
+  margin-top: 1rem;
+  padding-bottom: 1rem;
+  text-align: center;
+  margin-right: 500px;
+
+  @media ${device.tablet} {
+    text-align: left;
+    margin-right: 0;
+    margin-left: 30px;
+  }
+`;
+
+export const Brødsmule = styled.span`
+  margin-left: 5px;
+  margin-right: 5px;
+
+  a {
+    text-decoration: none;
+  }
+
+  a:hover {
+    text-decoration: underline;
+  }
+`;
 
 const Brødsmuler = () => {
   return (
-    <StyledBrødsmuler>
+    <BrødsmulerWrapper>
       <Brødsmule>
         <a href="https://nav.no">Forside</a>
       </Brødsmule>{' '}
@@ -12,7 +41,7 @@ const Brødsmuler = () => {
         <a href="https://www.nav.no/familie/alene-med-barn">Alene med barn</a>
       </Brødsmule>{' '}
       / <Brødsmule className="brødsmule-element">Hva kan du få?</Brødsmule>
-    </StyledBrødsmuler>
+    </BrødsmulerWrapper>
   );
 };
